Trait async_ffi::FutureExt

pub trait FutureExt: Future + Sized {
    // Provided methods
    fn into_ffi<'a>(self) -> BorrowingFfiFuture<'a, Self::Output> 
       where Self: Send + 'a { ... }
    fn into_local_ffi<'a>(self) -> LocalBorrowingFfiFuture<'a, Self::Output> 
       where Self: 'a { ... }
}

Helper trait to provide conversion from Future to FfiFuture or LocalFfiFuture.

See module level documentation for more details.

Provided Methods

fn into_ffi<'a>(self) -> BorrowingFfiFuture<'a, Self::Output>
where Self: Send + 'a,

Convert a Rust Future implementing Send into a FFI-compatible FfiFuture.

fn into_local_ffi<'a>(self) -> LocalBorrowingFfiFuture<'a, Self::Output>
where Self: 'a,

Convert a Rust Future into a FFI-compatible LocalFfiFuture.

Object Safety

This trait is not object safe.